+++ lib/common/common.types.ts
@@ -98,3 +98,8 @@ interface NestedQueries<T> {
 export interface GenericSearchFilters<T> {
     query: FlatQuery<T> | NestedQueries<T>;
 }
+
+export enum Order {
+    DESC = 'desc',
+    ASC = 'asc',
+}